# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/ping":{"get":{"summary":"Ping","operationId":"ping_ping_get","responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/home":{"get":{"summary":"Home","operationId":"home_home_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/posts":{"get":{"summary":"Get Posts","operationId":"get_posts_posts_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Post"},"type":"array","title":"Response Get Posts Posts Get"}}}}}},"put":{"summary":"Update Or Add Posts","operationId":"update_or_add_posts_posts_put","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Post"},"type":"array","title":"Updated Posts"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Post"},"type":"array","title":"Response Update Or Add Posts Posts Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Posts","operationId":"create_posts_posts_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Post"},"type":"array","title":"New Posts"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Post"},"type":"array","title":"Response Create Posts Posts Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ping/auth":{"get":{"summary":"Ping Auth","operationId":"ping_auth_ping_auth_get","responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Post":{"properties":{"author":{"type":"string","title":"Author"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"creation_datetime":{"type":"string","format":"date-time","title":"Creation Datetime"}},"type":"object","required":["author","title","content","creation_datetime"],"title":"Post"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}